+++ src/commonMain/kotlin/com/algolia/search/model/settings/Settings.kt
@@ -27,9 +27,11 @@ import com.algolia.search.serialize.KeyAttributesForFaceting
 import com.algolia.search.serialize.KeyAttributesToHighlight
 import com.algolia.search.serialize.KeyAttributesToRetrieve
 import com.algolia.search.serialize.KeyAttributesToSnippet
+import com.algolia.search.serialize.KeyAttributesToTransliterate
 import com.algolia.search.serialize.KeyCamelCaseAttributes
 import com.algolia.search.serialize.KeyCustomNormalization
 import com.algolia.search.serialize.KeyCustomRanking
+import com.algolia.search.serialize.KeyDecompoundQuery
 import com.algolia.search.serialize.KeyDecompoundedAttributes
 import com.algolia.search.serialize.KeyDisableExactOnAttributes
 import com.algolia.search.serialize.KeyDisablePrefixOnAttributes
@@ -454,6 +456,23 @@ public data class Settings(
      * lesser value - less relevant results.
      */
     @SerialName(KeyRelevancyStrictness) override var relevancyStrictness: Int? = null,
+
+    /**
+     * Enable word segmentation (also called decompounding) at query time for compatible languages. For example, this
+     * turns the Dutch query "spaanplaatbehang" into "spaan plaat behang" to retrieve more relevant results.
+     * Engine default: true
+     * [Documentation](https://www.algolia.com/doc/api-reference/api-parameters/decompoundQuery/?client=kotlin)
+     */
+    @SerialName(KeyDecompoundQuery) override var decompoundQuery: Boolean? = null,
+
+    /**
+     * Specify on which attributes to apply transliteration. Transliteration refers to the ability of finding results in
+     * a given alphabet with a query in another alphabet. For example, in Japanese, transliteration enables users to
+     * find results indexed in Kanji or Katakana with a query in Hiragana.
+     * Engine default: [*]
+     * [Documentation](https://www.algolia.com/doc/api-reference/api-parameters/attributesToTransliterate/?language=kotlin)
+     */
+    @SerialName(KeyAttributesToTransliterate) override var attributesToTransliterate: List<Attribute>? = null
 ) : SettingsParameters {
 
     /**