+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Unreleased
 
 ### Added
-- Query and Settings: `relevancyStrictness` parameter
+- Query/Settings: `relevancyStrictness` and `decompoundQuery` parameters (#248)
 - ResponseSearch: `appliedRelevancyStrictness` and `nbSortedHits` parameters
+- Settings: `attributesToTransliterate` parameter (#249)
 
 # 1.6.2